requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
cryptography==45.0.7
deep-translator==1.11.4
distro==1.9.0
filelock==3.19.1
firebase_admin==7.1.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.10
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.1
google-cloud-vision==3.10.2
google-crc32c==1.7.1
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
googletrans==4.0.0rc1
groq==0.31.1
grpc-stubs==1.53.0.6
grpcio==1.74.0
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httplib2==0.30.2
httpx==0.28.1
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
narwhals==2.4.0
networkx==3.5
numpy==2.3.3
openai==1.3.0
packaging==24.2
pandas==2.3.2
pillow==10.4.0
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.27.1
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SpeechRecognition==3.14.3
streamlit==1.37.0
streamlit_mic_recorder==0.0.8
sympy==1.14.0
tenacity==8.5.0
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
types-protobuf==4.25.0.20240417
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchdog==4.0.2
